I think I figured out how to initialize the User? Part of it is done in database_init, part in the User class (called from Action constructor).

+++ src/ResearchOS/database_init.py
@@ -5,21 +5,20 @@ import sys
 sys.path.append("/Users/mitchelltillman/Desktop/Not_Work/Code/Python_Projects/Biomech-Analysis-Platform-Python/src")
 import ResearchOS as ros
 from ResearchOS.action import Action
+from ResearchOS.research_object import DEFAULT_USER_ID
 
-# from ResearchOS.config import Config
-config = ros.Config()
-
-DEFAULT_USER_ID = "US000000_000"
 
 intended_tables = [
     "research_objects", "actions", "attributes", "research_object_attributes", "data_values", "data_addresses", "data_address_schemas", "current_user", "settings"
 ]
 
 class DBInitializer():
-    def __init__(self):        
+    def __init__(self, remove: bool = True):     
+        config = ros.Config()   
         db_file = config.db_file
         Action.conn.close()
-        if os.path.exists(db_file):
+        self.remove = remove
+        if os.path.exists(db_file) and self.remove:
             os.remove(db_file)
         Action.conn = sqlite3.connect(db_file)
         self._conn = Action.conn
@@ -31,9 +30,12 @@ class DBInitializer():
         self.create_database()
         self._conn.commit()
         self.check_tables_exist()
-        self.init_current_user_id()        
-        self._conn.cursor().close()
-        self._conn.close()
+        # self.init_current_user_id()
+        if not self.remove:
+            self._conn.cursor().close()
+            self._conn.close()
+            Action.conn = sqlite3.connect(db_file)
+            self._conn = Action.conn
 
     def check_tables_exist(self):
         """Check that all of the tables were created."""
@@ -57,6 +59,7 @@ class DBInitializer():
         cursor.execute(sqlquery)
         sqlquery = f"INSERT INTO actions (action_id, user_object_id, name, timestamp) VALUES ('{action_id}', '{user_id}', 'Initialize current user', '{datetime.datetime.now(datetime.UTC)}')"
         cursor.execute(sqlquery)
+        # Put the attributes into the research objects attributes table.
         self._conn.commit()             
 
     def create_database(self):
